Anchor text for topics changed to the info instead of the content
This largely arises from the complications of creating a link from a
full-body of text. If you embed an <a>nchor in your text I would need
to filter it out in order to make the text an <a>nchor itself.

I'm not about that + I think this way looks much nicer