XML for RSS feed properly escaped & some template bugfixes
I found the most beautiful thing in using the cast to an (array)
to convert an object into a usable assoc. array. This way I can avoid
generating StdObjects (fuck that) when I can simply pass in an assoc.
array of items that I want in the nav. God bless PHP