Remove robocheck CAPTCHA in favor of CAPTCHA-less robocheck
This is kind of a big deal; robocheck was (for the longest time) how
I limited spam on the site. Because of a rise in the number of posts which
are originating from *manual* submissions I have decided to remove the
image-based CAPTCHA from RAL in favor of a much simpler, 2-pronged system.

1. Honeypot <form> fields
2. Anti-spam filtering using a Bayesian filter analysis

This commit addresses the first bullet point, and my next work
will aim towards implementing the second point.